requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cloudinary==1.44.1
django==5.2.6
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-prose-editor==0.18.3
django-taggit==6.1.0
django-unfold==0.65.0
djangorestframework==3.16.1
gunicorn==25.1.0
nh3==0.3.0
packaging==26.0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0
